iomap: move buffered io bio logic into new file
mainline inclusion
from mainline-v6.19-rc1
category: feature

Move bio logic in the buffered io code into its own file and remove
CONFIG_BLOCK gating for iomap read/readahead.

[1] https://lore.kernel.org/linux-fsdevel/aMK2GuumUf93ep99@infradead.org/

(cherry picked from commit c2b1adc4624573f9d5a35ccb8ebe31eb1bdb0d22)
